+++ esbuild-buffer-shim.js
@@ -1,0 +1,2 @@
+import { Buffer } from "buffer-es6";
+globalThis.Buffer = Buffer;